requirements.txt
streamlit
langchain
langchain-openai
langchain-pinecone
pinecone-client
openai
python-dotenv